LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/contain-intrinsic-size/auto-003-expected.txt


FAIL contain-intrinsic-size: auto assert_equals: expected 1 but got 0
